requirements.txt
Flask==3.0.3
Flask-Mail==0.10.0
Werkzeug==3.0.3
requests==2.31.0
python-dotenv==1.0.1
insightface==0.7.3
onnxruntime==1.17.3
opencv-python-headless==4.9.0.80
numpy==1.26.4
openpyxl==3.1.2
Pillow==10.3.0
gunicorn==21.2.0
